Add snake app, migrate to new model and db table
